+++ CHANGELOG.md
@@ -13,7 +13,8 @@ Features:
 - Add `k-FormComposer` component.
 - Add new `FormAmountAndCurrency` React component.
 - Add new `FormPhoneNumber` React component.
-- New `DevGrid` React component.
+- New `DevGrid` React component and new CSS components `k-DevGrid` and
+  `k-DevBreakpoint`.
 
 Fixes:
 - Change font color for disabled elements on checkboxes and radio buttons.